refactor(milli): move index_documents tests to a separate file
`index_documents/mod.rs` carried close to 2900 lines of tests at the end
of the file, which slows down rust-analyzer and makes the module harder
to navigate.

Move the whole `#[cfg(test)] mod tests` block to a sibling
`mod_test.rs` file, included via `#[cfg(test)] mod mod_test;`, following
the convention already used elsewhere in the codebase (e.g.
`search/mod_test.rs`, `option_test.rs`). No test logic is changed.

Part of #5116
